# Optimize available quantity loader. 