@patrys 2021-08-11
Task linked: [SALEOR-4020 Long-running queries by dataloader AvailableQuantityByProductVariantIdCountryCodeAndChannelSlugLoader](https://app.clickup.com/t/2549495/SALEOR-4020)